Remove suspicious type cast

+++ check_interfaces.c
@@ -377,15 +377,18 @@ int main(int argc, char *argv[]) {
 				if (vars->type == ASN_OCTET_STR) {
 					if (config.trimdescr && config.trimdescr < vars->val_len) {
 						interfaces[count].index =
-							(int)vars->name[(vars->name_length - 1)];
+							vars->name[(vars->name_length - 1)];
+
 						MEMCPY(interfaces[count].descr,
 							   (vars->val.string) + config.trimdescr,
 							   vars->val_len - config.trimdescr);
+
 						TERMSTR(interfaces[count].descr,
 								vars->val_len - config.trimdescr);
 					} else {
 						interfaces[count].index =
-							(int)vars->name[(vars->name_length - 1)];
+							vars->name[(vars->name_length - 1)];
+
 						MEMCPY(interfaces[count].descr, vars->val.string,
 							   vars->val_len);
 						TERMSTR(interfaces[count].descr, vars->val_len);
